zcash_pool_migration: split `testing` into strategies, conformance, scenarios
The module held three things whose consumers are disjoint: `proptest`
generators, the backend-agnostic conformance suites, and the fixed golden
data. A downstream store crate wants the suites, the end-to-end tests want
the scenarios, and the codec proptests want the generators, so a reader
looking for any one of them waded through the other two.

Each group is now its own submodule, re-exported from `testing/mod.rs`, so
every existing path resolves unchanged: `git diff origin/main` touches no
consumer at all.

The split also makes the layering visible. `conformance` and `strategies`
both depend on `scenarios` and not on each other, which says what these
things are: the scenarios are the data, the strategies generate more of it,
and the conformance suites are the satisfaction relation over both.
